requirements.txt
fastapi
uvicorn
openai
pytest
pytest-asyncio
minio
python-magic
